Remove the Implementation trait
Its main gain over simple FnMut(..)s was the downcasting abilities, to
check if an object has a given implementation. This advantage is no
longer present now that we have typed user_data since #198.

As such, lets get rid of Implementation, for the sake of simplicity.

To fully retreive all functionnality, it is needed to also introduce an
user-data mechanism on event sources, but this will be done when #185 is
adressed.

Fixes #197.