server: Only filter rust-managed globals
When interfacing with a C library (like MESA) which creates its own
globals, their user_data won't be setup the way our global filter
expects. As a result, we must not attempt to filter them.

This achieves it by keeping a list of globals we manage, and ignoring
unknown globals in the filtering process.

This means that currently, C-created globals cannot be filtered.