[client] Introduce Proxy::send_constructor
This new method of `Proxy` sends a message that is expected to create an
object, returning the newly created object. The input message's slot for
the object to create must have previously been filled with a
"placeholder" object (this is checked and failure to do so will panic).

This allows to avoid a race when two threads would conccurently create
objects to send and they end up sent to the server in the wrong order,
causing a protocol error.

fixes #206