client+server: make non-send impls the default
The implementation methods are now `.implement()` which allows
non-`Send` implementations and panics if called from the wrong
thread, and `.implement_threadsafe()` which accepts only `Send`
implementations and can be called from any thread.

This is accompanied by removing any `Send` / `Sync` implementations
server-side when the `native_lib` feature is activated, as the
C library actually is never threadsafe (contrarily to the client
one).