client: Summon dead Proxies for dead arguments
When we receive an event with a non-nullable object argument which
points to a non-existing object, summon a dead object to replace it
rather than crashing.

This can happen if we destroyed the target object conccurently with the
server sending us this event, and we should not crash due to that.

Fixes #334