wayland-commons: avoid empty sendmsg on flush
If there is nothing to send, there's no point actually making the system call.
The list of FDs is not checked because sending FDs without also sending data is
not permitted anyway.

This also updates tests to detect errors on reads (not writes), which is where
a real client would expect to detect them.  The native client requires two
reads because the first one will just read the protocol error message.