Switch default socket name to wayland-1
It has come to my attention a while ago that it seems to be good pratice
now to use `wayland-1` as the default socket name rather than
`wayland-0`.

This should prevent clients from connecting to `wayland-0` even if
`WAYLAND_DISPLAY` is not set, which is why this change should be
exclusive to the wayland-server part.