Add Proxy::send_{request,constructor}() for manual dispatch
This allows sending a Request that creates an object using a custom
ObjectData to handle events instead of using the Dispatch mechanism.

The two functions are split in order to avoid needing to explicitly
specify a generic type when not creating a child (since it will be
impossible to infer in that case).