backend/scanner: Fix handling of nullable strings
A `NULL` value in a nullable string is distict from an empty non-`NULL`
string. So this changes `Argument::Str` to contain an `Option`, updates
`sys` to match how this is represented in libwayland (as a `NULL`
pointer), updates `rs` to match the wire protocol for this, and fixes
unit tests and `wayland-scanner` accordingly.

Fixes segfault when `NULL` was passed to `wl_data_offer::accept` with
the `sys` backend.

I was going to make changes for nullable arrays too, but those appear to
be broken in libwayland as well, so that can stay as is for now.
https://gitlab.freedesktop.org/wayland/wayland/-/issues/306