client: Add a GTK example
This provides an example of how to use `wayland_client` with a foreign
display, in particular GTK. If there is a better way to use
`wayland_client` with GTK, the example can be improved.

I'm not sure if this use of `prepare_read` is correct. I guess it can't
really run into issues given no other thread is actually reading from
the connection? Should it instead create a guard before calling
`unix_fd_add_local`, and replace it with a new guard every time the
function is called?