Replace io-lifetimes with std::os::fd
Since Rust 1.63, most of the features of io-lifetimes have been merged
into std, including the only three we are using: OwnedFd, BorrowedFd and
the AsFd trait.

This bumps our MSRV, but this is the MSRV of the newest io-lifetimes
anyway.

I kept reexporting io_lifetimes as it was part of the public interface,
but since it isn’t used any more it would be nice to remove it before
the next minor version bump.  I can prepare a commit for that.