Log requests on dead objects
To make it possible to understand that something got discarded,
print a message saying so inside when `WAYLAND_DEBUG` is enabled.

This also changes how timestamps are computed for `WAYLAND_DEBUG` log
so they align with `libwayland-client` to look into logs using both
at the same time more accessible. The `[rs]` particle was added to
the log to give a hint which library was used, which could
be helpful when looking into random logs.