client: Include version info and interface name in `BindError`
When `BindError` is returned (and bubbled up by callers) it is no longer
clear what interface or version was requested (nor what is the highest
this server might be able to provide), leaving users to wonder what the
context for the error is.

Add this information to the object to elaborate in error messages and
allow callers to read the values after the fact.