Use a wrapper type implementing `Sync`, instead of `static mut` (#700)
* Use a wrapper type implementing `Sync`, instead of `static mut`

`static mut` is generally best avoided. It seems to only be necessary
currently to allow the use of non-`Sync` types, since pointers are not
sync.

This also removes the `NULLPTR` constant in favor of using
`std::ptr::null`, uses `.as_ptr()` instead of casting, and uses the
arrray repeating syntax to define `types_null`. None of those should
impact behavior or the public API.

`*_requests` and `*_events` static are no longer public now. This is
theoretically a breaking change, but shouldn't really impact anything.

* Suppress `clippy::test_attr_in_doctest` lint

* Address nightly warnings