backend/client_sys: add read_without_dispatch
Add a way to read the event queue, but don't do an immediate dispatch
of the events, leaving it up to dispatch_inner_queue. This is useful,
when we spawn a separate thread that interacts with the main queue of
the backend, like for example to track if events generally arrived and
we have something to dispatch.