tests: Add test for error when server ring buffer is full
Should only one error be returned here (on both backends)? For now,
check for both connection reset and broken pipe errors.

When we add a way to set the buffer size, we can have more tests here,
to verify first that an error occurs without increasing the buffer size,
then that it works without error after increasing the size as needed.